moved config.json to root and named it lune-config
