+++ lune-config.json
@@ -1,5 +1,5 @@
 {
-	"$schema": "../config-schema.json",
+	"$schema": "config-schema.json",
 	"lunePort": 6548,
 	"mainScript": "src/init.luau",
 	"executable": "src/init.exe"
